Create a pan and zoom image component
