+++ src/vs/workbench/contrib/positronPlots/browser/components/staticPlotInstance.tsx
@@ -3,6 +3,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as React from 'react';
+import { PanZoomImage } from 'vs/workbench/contrib/positronPlots/browser/components/panZoomImage';
 import { ZoomLevel } from 'vs/workbench/contrib/positronPlots/browser/components/zoomPlotMenuButton';
 import { StaticPlotClient } from 'vs/workbench/services/positronPlots/common/staticPlotClient';
 
@@ -25,114 +26,12 @@ interface StaticPlotInstanceProps {
  * @returns The rendered component.
  */
 export const StaticPlotInstance = (props: StaticPlotInstanceProps) => {
-	const [width, setWidth] = React.useState<number>(1);
-	const [height, setHeight] = React.useState<number>(1);
-	const imageWrapperRef = React.useRef<HTMLDivElement>(null);
-	const [moveX, setMoveX] = React.useState<number>(0);
-	const [moveY, setMoveY] = React.useState<number>(0);
-	const [grabbing, setGrabbing] = React.useState<boolean>(false);
-
-	const onImgLoad = (event: React.SyntheticEvent<HTMLImageElement>) => {
-		const img = event.target as HTMLImageElement;
-		const width = img.naturalWidth;
-		const height = img.naturalHeight;
-		setWidth(width);
-		setHeight(height);
-	};
-
-	React.useEffect(() => {
-		const { clientWidth, clientHeight } = imageWrapperRef.current ?? { clientWidth: 0, clientHeight: 0 };
-		setMoveX((clientWidth - (width * props.zoom)) / 2);
-		setMoveY((clientHeight - (height * props.zoom)) / 2);
-	}, [width, height, props.zoom]);
-
-	React.useLayoutEffect(() => {
-		if (!imageWrapperRef.current || props.zoom === ZoomLevel.Fill) {
-			return;
-		}
-	});
-
-	const getStyle = (): React.CSSProperties => {
-		const panMovement = `translateX(${moveX}px) translateY(${moveY}px)`;
-		let style: React.CSSProperties = {};
-
-		if (props.zoom === ZoomLevel.Fill) {
-			// no centering and let the entire plot be visible
-			style = {
-				width: '100%',
-				height: '100%',
-				position: 'unset',
-				transform: 'none',
-			};
-		} else {
-			style = {
-				cursor: grabbing ? 'grabbing' : 'grab',
-				maxWidth: 'none',
-				maxHeight: 'none',
-				top: 0,
-				left: 0,
-				transform: panMovement,
-			};
-		}
-
-		return style;
-	};
-
-	const applyZoom = (value: number): string => {
-		if (props.zoom !== ZoomLevel.Fill) {
-			return `${value * props.zoom}px`;
-		}
-
-		return '100%';
-	};
-
-	const panImage = (event: React.MouseEvent<HTMLImageElement>) => {
-		const { clientWidth, clientHeight } = imageWrapperRef.current ?? { clientWidth: 0, clientHeight: 0 };
-		const adjustedWidth = width * props.zoom;
-		const adjustedHeight = height * props.zoom;
-		const maxMoveX = adjustedWidth > clientWidth ? 0 : clientWidth - adjustedWidth;
-		const minMoveX = adjustedWidth > clientWidth ? -adjustedWidth + clientWidth : 0;
-		const maxMoveY = adjustedHeight > clientHeight ? 0 : clientHeight - adjustedHeight;
-		const minMoveY = adjustedHeight > clientHeight ? -adjustedHeight + clientHeight : 0;
-
-		let newMoveX = 0, newMoveY = 0;
-
-		if (event.type === 'wheel' && event.buttons === 0) {
-			const wheelEvent = event as React.WheelEvent<HTMLImageElement>;
-			newMoveX = moveX + wheelEvent.deltaX;
-			newMoveY = moveY + wheelEvent.deltaY;
-		} else if (event.buttons === 1 && props.zoom !== ZoomLevel.Fill) {
-			newMoveX = moveX + event.movementX;
-			newMoveY = moveY + event.movementY;
-		} else {
-			return;
-		}
-
-		const finalMoveX = Math.max(Math.min(newMoveX, maxMoveX), minMoveX);
-		const finalMoveY = Math.max(Math.min(newMoveY, maxMoveY), minMoveY);
-		setMoveX(finalMoveX);
-		setMoveY(finalMoveY);
-	};
-
-	const updateCursor = (event: React.MouseEvent<HTMLImageElement>) => {
-		setGrabbing(event.type === 'mousedown');
-	};
-
 	return (
 		<div className='plot-instance static-plot-instance'>
-			<div className='image-wrapper' ref={imageWrapperRef}>
-				<img src={props.plotClient.uri}
-					alt={props.plotClient.code ? props.plotClient.code : 'Plot ' + props.plotClient.id}
-					onLoad={onImgLoad}
-					style={getStyle()}
-					width={applyZoom(width)}
-					height={applyZoom(height)}
-					onMouseMoveCapture={panImage}
-					onMouseDown={updateCursor}
-					onMouseUp={updateCursor}
-					onWheel={panImage}
-					draggable={false}
-				/>
-			</div>
+			<PanZoomImage
+				imageUri={props.plotClient.uri}
+				description={props.plotClient.code ? props.plotClient.code : 'Plot ' + props.plotClient.id}
+				zoom={props.zoom}
+			/>
 		</div>);
 };
